Split database schema to a new file (#140)
We are decoupling the database from history to be a new schema by
itself.
