+++ data/release/xdg/config.toml
@@ -9,15 +9,9 @@ file = "/tmp/command-line-assistant/output.txt"
 prompt_separator = "$"
 # ------------------ experimental
 
-# History management configuration
-[history]
-# If the history for all conversation should be enabled or not. By default,
-# this is enabled and will store in a database defined under
-# [history.database]. The conversation is stored per-user.
-enabled = true
 
 # History Database settings. By default, sqlite is used.
-[history.database]
+[database]
 # Available types for databases are: sqlite, postgresql and mysql. Mariadb can
 # be used as well with "mysql" type.
 type = "sqlite"
@@ -26,7 +20,7 @@ type = "sqlite"
 connection_string = "/var/lib/command-line-assistant/history.db"
 
 # In order to use postgresql, uncomment the following settings:
-# [history.database]
+# [database]
 # type = "postgresql"
 # host = "localhost"
 # port = "5432"
@@ -35,13 +29,20 @@ connection_string = "/var/lib/command-line-assistant/history.db"
 # password = "your-password"
 
 # Or, to use mysql, uncomment the following:
-# [history.database]
+# [database]
 # type = "mysql"
 # host = "localhost"
 # port = "3306"
 # database = "history"
 # user = "your-user"
 # password = "your-password"
+
+# History management configuration
+[history]
+# If the history for all conversation should be enabled or not. By default,
+# this is enabled and will store in a database defined under
+# [database]. The conversation is stored per-user.
+enabled = true
 
 # Backend settings for communicating with the external API.
 [backend]
@@ -67,7 +68,7 @@ level = "INFO"
 # Global setting to enable logging responses in an audit file and journald
 responses = true
 # Global setting to enable logging question in an audit file and journald
-question = true  # Global setting - don't log questions by default
+question = true # Global setting - don't log questions by default
 
 # User-specific settings that can override the global settings based on user name.
 # Note: The user must exist in order for this to work.
