--- data/release/xdg/history.json
@@ -1,1 +1,0 @@
-{"history": [],"metadata": {"last_updated": "2024-12-30T12:04:36.298732Z",  "version": "0.1.0",  "entry_count": 0,  "size_bytes": 0}}
