+++ docs/source/config/schemas/database.rst
@@ -1,0 +1,8 @@
+Database
+========
+
+.. automodule:: command_line_assistant.config.schemas.database
+   :members:
+   :undoc-members:
+   :private-members:
+   :no-index:
