+++ tests/config/schemas/test_schema.py
@@ -1,7 +1,8 @@
 import pytest
 
 from command_line_assistant.config.schemas.backend import AuthSchema, BackendSchema
-from command_line_assistant.config.schemas.history import DatabaseSchema, HistorySchema
+from command_line_assistant.config.schemas.database import DatabaseSchema
+from command_line_assistant.config.schemas.history import HistorySchema
 from command_line_assistant.config.schemas.logging import LoggingSchema
 from command_line_assistant.config.schemas.output import OutputSchema
