+++ tests/conftest.py
@@ -7,7 +7,8 @@ import pytest
 
 from command_line_assistant import config, logger
 from command_line_assistant.config.schemas.backend import AuthSchema, BackendSchema
-from command_line_assistant.config.schemas.history import DatabaseSchema, HistorySchema
+from command_line_assistant.config.schemas.database import DatabaseSchema
+from command_line_assistant.config.schemas.history import HistorySchema
 from command_line_assistant.config.schemas.logging import LoggingSchema
 from command_line_assistant.config.schemas.output import OutputSchema
 from command_line_assistant.dbus.context import DaemonContext
@@ -67,8 +68,8 @@ def mock_config(tmp_path):
             ),
             history=HistorySchema(
                 enabled=True,
-                database=DatabaseSchema(type="sqlite", connection_string=history_db),
             ),
+            database=DatabaseSchema(type="sqlite", connection_string=history_db),
             logging=LoggingSchema(
                 level="debug",
                 users={"testuser": {"question": True, "responses": False}},
